pyproject.toml
[tool.poetry]
name = "costpi-backend"
version = "0.1.0"
description = "Backend Django do CostPI"
authors = ["zPaulo <paulotjs147@hotmail.com>"]
readme = "README.md"
packages = [{include = "apps", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.12"
django-ninja = ">=1.4.1,<2.0.0"
django-extensions = ">=4.1,<5.0"
python-decouple = ">=3.8,<4.0"
psycopg2-binary = ">=2.9.10,<3.0.0"
gunicorn = "^23.0.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.7"

[tool.ruff]
line-length = 88
target-version = "py312"
exclude = ["migrations", ".venv"]

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]
fixable = ["ALL"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
